backend/requirements/base.txt
# Project's base packages
djangorestframework==3.14.0
Django==5.1.1
Markdown==3.4.4
django-filter==23.2
django-environ==0.10.0
django-cors-headers==4.5.0
# DRF-spectacular for api documentation
drf-spectacular==0.26.4
# Developer's tools
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
requests==2.32.3
pillow==11.0.0
qrcode==8.0
aiohttp==3.11.8
docxtpl==0.19.0
httpx==0.28.1
ruff==0.12.2